pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "search_names"
version = "0.4.0"
description = "Advanced name search and entity recognition in large text corpora with NLP and AI capabilities"
authors = [
    {name = "Suriyan Laohaprapanon", email = "suriyant@gmail.com"},
    {name = "Gaurav Sood", email = "gsood07@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities"
]
keywords = ["search", "names", "text", "corpus", "pattern", "ner", "nlp", "entity-recognition", "entity-linking", "fuzzy-matching"]
dependencies = [
    "nameparser",
    "python-Levenshtein",
    "regex",
    "nltk",
    "typer[all]>=0.9.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = ["ruff", "black", "isort", "mypy", "types-PyYAML"]
test = ["coverage", "pytest", "pytest-asyncio"]
docs = ["sphinx", "furo"]
nlp = [
    "spacy>=3.7.0",
    "sentence-transformers>=2.2.0",
    "transformers>=4.21.0",
]
ml = [
    "parsernaam>=0.0.1",  # ML-based name parsing for complex names
]
formats = [
    "pyarrow>=10.0.0",  # For Parquet support
    "polars>=0.19.0",   # Modern DataFrame library
]
search = [
    "elasticsearch>=8.0.0",
    "opensearch-py>=2.0.0",
]
web = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
]
all = [
    "search_names[nlp,ml,formats,search,web]"
]

[project.urls]
Homepage = "https://github.com/appeler/search_names"
Repository = "https://github.com/appeler/search_names"
Issues = "https://github.com/appeler/search_names/issues"

[project.scripts]
# Modern CLI - primary interface
search-names = "search_names.cli:main_cli"

# Legacy CLI commands - for backward compatibility
clean_names = "search_names.clean_names:main"
merge_supp = "search_names.merge_supp:main"
preprocess = "search_names.preprocess:main"
split_text_corpus = "search_names.split_text_corpus:main"
search_names = "search_names.search_names:main"
merge_results = "search_names.merge_results:main"

[tool.hatch.build.targets.wheel]
packages = ["search_names"]

[tool.hatch.build.targets.wheel.shared-data]
"examples" = "share/search_names/examples"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by formatter
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E402"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "nameparser.*",
    "parsernaam.*",
    "Levenshtein.*",
    "regex.*",
    "nltk.*",
]
ignore_missing_imports = true